requirements.txt
django==4.2
djangorestframework==3.14.0
django-environ==0.10.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
dj-rest-auth==4.0.0
psycopg2-binary==2.9.6
pytest==7.3.1
Faker==18.7.0
factory-boy==3.2.1
Pillow==9.5.0